sender email change bigent

+++ database/migrations/2024_09_10_092403_create_mailsetup_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->string('sender_department')->nullable();
             $table->string('sender_company_logo')->nullable();
             $table->string('sender_website')->nullable();
-            $table->integer('sender_number')->nullable();
+            $table->unsignedBigInteger('sender_number')->nullable();
             $table->json('other_links');
             $table->foreignId('user_id')->references('id')->on('users')->cascadeOnDelete();
             $table->timestamps();
